Fix implementation bugs weakdes_gfp.cc
BUG 1:
The indices are computed with bit shifting rather than using the
theoretical formula, i.e.,

  index_i(a, p_i(a)) = a + t * p_i(a) .

Bit shifting and above formula are equal only when using extended Galois
fields with powers of 2, not for any prime number t.

BUG 2:
The coefficients are also computed with with bit shifting rather than
using the theoretical formula from the paper. Results only coincide when
using powers of 2, not for any prime number t.

BUG 3:
The polynomial evaluation is not done correctly using Horner's method
because the coefficients are taken in the inverse order. In practice,
this means that the basic construction from the paper is not followed.
This probably does not affect the correctness of the weak design as
previous two bugs.